refactor: 포커스 필드 로직을 카드 등록 훅으로 이동
